Adapt Cypress tests for OSSMC list pages
Add @ossmc tags to feature files for namespaces, istio_config, overview,
sidecar_injection, waypoint, and wizard suites so they run in the OSSMC
plugin context. Extract linkSelector, hasAtLeastOneClass, and
getClusterForSingleCluster into a shared utils.ts to decouple
non-table utilities from table.ts. Update link selectors in apps.ts,
istio_config.ts, services.ts, and namespaces.ts to match both
<a href> and <button data-href> for OSSMC kiosk mode. Fix cy.visit
calls in navigation.ts to use object format and glob-based intercept
patterns for proxied API requests.

Ref: kiali/openshift-servicemesh-plugin#664
Made-with: Cursor
