+++ frontend/cypress/integration/featureFiles/istio_config_validation_filters.feature
@@ -1,4 +1,5 @@
 @istio-config-validation
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Kiali Istio Config page
