+++ frontend/cypress/integration/featureFiles/namespaces.feature
@@ -1,4 +1,5 @@
 @namespaces
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Kiali Namespaces page
