+++ frontend/cypress/integration/featureFiles/overview.feature
@@ -1,4 +1,5 @@
 @overview
+@ossmc
 Feature: New Overview - Overview cards
 
   Background:
