+++ frontend/cypress/integration/featureFiles/sidecar_injection.feature
@@ -1,4 +1,5 @@
 @sidecar-injection
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Controlling sidecar injection
