+++ frontend/cypress/integration/featureFiles/waypoint.feature
@@ -1,4 +1,5 @@
 @waypoint
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 # TODO: offline - ambient support.
 Feature: Kiali Waypoint related features
