+++ frontend/cypress/integration/featureFiles/wizard_istio_config.feature
@@ -1,4 +1,5 @@
 @wizard-istio-config
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Kiali Istio Config wizard
