+++ frontend/cypress/integration/featureFiles/wizard_k8sgw_api_grpc_routing.feature
@@ -1,4 +1,5 @@
 @wizard-k8sgw-api-routing
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Service Details Wizard: K8s GRPC Routing
