+++ frontend/cypress/integration/featureFiles/wizard_request_routing.feature
@@ -1,4 +1,5 @@
 @wizard-request-routing
+@ossmc
 # don't change first line of this file - the tag is used for the test scripts to identify the test suite
 
 Feature: Service Details Wizard: Request Routing
